ShipDock moved to SetupMapScreen file
